+++ src/dec_cutpacking.c
@@ -377,7 +377,7 @@ SCIP_RETCODE buildNewAdjacencyList(
       nconss = j;
 
       /* insert representative */
-      SCIP_CALL( SCIPhashmapInsert(detectordata->representatives, (void*) (size_t) (detectordata->nrepresentatives + 1), representative ) );
+      SCIP_CALL( SCIPhashmapInsert(detectordata->representatives, (void*) ((size_t) detectordata->nrepresentatives + 1), representative ) );
       SCIP_CALL( copyhashmap(consslink,detectordata->mergedconss[detectordata->nrepresentatives]) );
       detectordata->nrepresentatives++;
 
@@ -630,7 +630,7 @@ SCIP_RETCODE buildNewGraphs(
    pos2 = -1;
    for( i = 0; i < detectordata->nrelconss + 1; ++i )
    {
-      if( (!SCIPhashmapExists(detectordata->occupied, (void*)(size_t)(i + 1))) )
+      if( (!SCIPhashmapExists(detectordata->occupied, (void*) ((size_t)i + 1))) )
       {
          if( pos1 == -1 )
             pos1 = i;
@@ -646,7 +646,7 @@ SCIP_RETCODE buildNewGraphs(
 
    SCIP_CALL( SCIPallocMemoryArray(scip, &detectordata->graphs[pos1].conss, graph.nconss) );
    SCIP_CALL( SCIPallocMemoryArray(scip, &detectordata->graphs[pos2].conss, graph.nconss) );
-   SCIP_CALL( SCIPhashmapRemove(detectordata->occupied, (void*)(size_t)(detectordata->position + 1)) );
+   SCIP_CALL( SCIPhashmapRemove(detectordata->occupied, (void*) ((size_t)detectordata->position + 1)) );
 
    for( i = 0; i < graph.nconss; ++i )
    {
@@ -799,20 +799,20 @@ SCIP_RETCODE buildNewGraphs(
    {
       SCIP_CALL( copyConss(scip, detectordata, pos1, nconss1) );
       SCIP_CALL( setStartBlock(scip, detectordata,detectordata->graphs[pos1].cons1) );
-      SCIP_CALL( SCIPhashmapInsert(detectordata->occupied, (void*) (size_t) (pos2 + 1), NULL) );
+      SCIP_CALL( SCIPhashmapInsert(detectordata->occupied, (void*) ((size_t)pos2 + 1), NULL) );
       SCIP_CALL( freeGraph(scip, detectordata, pos1, nconss1) );
    }
    else if( (nconss2 < 2) || ((stop1 == 0) && stop2) )
    {
       SCIP_CALL( copyConss(scip, detectordata, pos2, nconss2) );
       SCIP_CALL( setStartBlock(scip, detectordata,detectordata->graphs[pos2].cons1) );
-      SCIP_CALL( SCIPhashmapInsert(detectordata->occupied, (void*) (size_t) (pos1 + 1), NULL) );
+      SCIP_CALL( SCIPhashmapInsert(detectordata->occupied, (void*) ((size_t)pos1 + 1), NULL) );
       SCIP_CALL( freeGraph(scip, detectordata, pos2, nconss2) );
    }
    else
    {
-      SCIP_CALL( SCIPhashmapInsert(detectordata->occupied, (void*) (size_t) (pos1 + 1), NULL) );
-      SCIP_CALL( SCIPhashmapInsert(detectordata->occupied, (void*) (size_t) (pos2 + 1), NULL) );
+      SCIP_CALL( SCIPhashmapInsert(detectordata->occupied, (void*) ((size_t)pos1 + 1), NULL) );
+      SCIP_CALL( SCIPhashmapInsert(detectordata->occupied, (void*) ((size_t)pos2 + 1), NULL) );
       detectordata->ngraphs++;
    }
 
@@ -847,7 +847,7 @@ SCIP_RETCODE getMergedConss(
    {
       for( j = 0; j < detectordata->nsubscipconss[i]; ++j )
       {
-         SCIP_CALL( hashmapInsert(constoblock, subscipconss[i][j], (void *) (size_t) (i+1)) );
+         SCIP_CALL( hashmapInsert(constoblock, subscipconss[i][j], (void *) ((size_t)i+1)) );
       }
    }
 
@@ -996,7 +996,7 @@ SCIP_RETCODE getConsIndex(
    {
       for( j = 0; j < nnewsubscipconss[i]; ++j )
       {
-         SCIP_CALL( SCIPhashmapSetImage(detectordata->constoblock, newsubscipconss[i][j], (void*) (size_t) (i+1)) );
+         SCIP_CALL( SCIPhashmapSetImage(detectordata->constoblock, newsubscipconss[i][j], (void*) ((size_t)i+1)) );
       }
    }
 
@@ -1093,7 +1093,7 @@ SCIP_RETCODE getLinkingVars(
       {
          linkingvars[nlinkingvars] = vars[i];
          ++nlinkingvars;
-         SCIP_CALL( SCIPhashmapInsert(vartoblock, vars[i], (void*) (size_t) (detectordata->nblocks+1)) );
+         SCIP_CALL( SCIPhashmapInsert(vartoblock, vars[i], (void*) ((size_t)detectordata->nblocks+1)) );
       }
       else
       {
@@ -1759,7 +1759,7 @@ DEC_DECL_INITDETECTOR(initCutpacking)
    SCIPreallocMemoryArray(scip, &(detectordata->relvars), j); /*lint !e522*/
 
    /* get number of relevant conss */
-   SCIP_CALL( SCIPallocMemoryArray(scip, &detectordata->graphs, nconss+1) );
+   SCIP_CALL( SCIPallocMemoryArray(scip, &detectordata->graphs, (size_t)nconss+1) );
    SCIP_CALL( SCIPallocMemoryArray(scip, &newconss, nconss) ); /*lint !e522*/
    k = 0;
    for( i = 0; i < nconss; ++i )
@@ -1924,7 +1924,7 @@ DEC_DECL_DETECTSTRUCTURE(detectAndBuildCutpacking)
    {
       for( i = 0; i < detectordata->nrelconss + 1; i++ )
       {
-         if( SCIPhashmapExists(detectordata->occupied, (void*) (size_t) (i + 1)) )
+         if( SCIPhashmapExists(detectordata->occupied, (void*) ((size_t)i + 1)) )
          {
             detectordata->position = i;
